+++ src/FluentHub/App.xaml.cs
@@ -29,7 +29,6 @@ using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
-
 namespace FluentHub
 {
     sealed partial class App : Windows.UI.Xaml.Application
