+++ src/FluentHub/UserControls/Labels/LabelControl.xaml.cs
@@ -15,7 +15,6 @@ using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
-
 namespace FluentHub.UserControls.Labels
 {
     public sealed partial class LabelControl : UserControl
