+++ src/FluentHub/ViewModels/MainViewModel.cs
@@ -1,4 +1,4 @@
-﻿using FluentHub.UserControls;
+﻿using FluentHub.Models;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
