+++ src/FluentHub/ViewModels/UserControls/ButtonBlocks/RepoButtonBlockViewModel.cs
@@ -17,7 +17,6 @@ namespace FluentHub.ViewModels.UserControls.ButtonBlocks
 
         public Repository Repository { get; set; }
 
-
         public event PropertyChangedEventHandler PropertyChanged;
         protected bool SetProperty<T>(ref T field, T newValue, [CallerMemberName] string propertyName = null)
         {
