+++ src/FluentHub/Views/MainPage.xaml.cs
@@ -16,7 +16,6 @@ using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
-
 namespace FluentHub.Views
 {
     public sealed partial class MainPage : Page
