+++ src/FluentHub/Views/Users/PullRequestListPage.xaml.cs
@@ -13,7 +13,6 @@ using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
-
 namespace FluentHub.Views.Users
 {
     public sealed partial class PullRequestListPage : Page
